+++ IHA/src/cz/vutbr/fit/iha/adapter/device/units/BlankUnit.java
@@ -23,7 +23,7 @@ public class BlankUnit extends BaseUnit {
 	}
 
 	@Override
-	public float convertValue(Item to, float value) {
+	public double convertValue(Item to, double value) {
 		return value;
 	}
