+++ IHA/src/cz/vutbr/fit/iha/adapter/device/units/TemperatureUnit.java
@@ -28,12 +28,12 @@ public class TemperatureUnit extends BaseUnit {
 	}
 	
 	@Override
-	public float convertValue(Item to, float value) {
+	public double convertValue(Item to, double value) {
 		switch (to.getId()) {
 		case FAHRENHEIT:
 			return value * 9 / 5 + 32;
 		case KELVIN:
-			return value + 273.15f;
+			return value + 273.15;
 		default:
 			return value;
 		}
