+++ IHA/src/cz/vutbr/fit/iha/adapter/device/units/UnknownUnit.java
@@ -23,7 +23,7 @@ public class UnknownUnit extends BaseUnit {
 	}
 
 	@Override
-	public float convertValue(Item to, float value) {
+	public double convertValue(Item to, double value) {
 		return value;
 	}
